requirements.txt
pyTelegramBotAPI
psycopg2-binary
python-dotenv
openpyxl
